Updated a lot and now it works
Fixed the get_gaze_coordinates with the new csv file data. So now the coordinates match the ones that map the red dot like in pupil capture. Now it's time to see if we can either split the world camera or maybe strap another camera to the classes to act as the live feed camera.