IT WORKS
Added a few things to help test but It works pretty well. Real-time works but only if you have another web cam plugged in to act as the other camera, because we cant split the camera to be in 2 places at once. Going to now work on building a dataset that could increase tesseracts accuracy. 2 weeks left its grind time. The Read me will be finalized soon